first measurement overhead benchmark add
